parse EDNS0 request packets with OPT RR trailer
Modern versions of BIND 9 dig send requests with a trailing
EDNS0-compliant (RFC 6891) pseudo-RR record, which trips up
the current approach to finding the QTYPE and QCLASS fields
by looking at the last 4 bytes of the request packet.

Because of this, running "dig foo.localhost" now typically
results in a failure (an NXDOMAIN reply), and the current
test suite mostly fails, reflecting this fact.

While we could fix the test suite by just adding +noedns to
all the dig queries, that would still leave the naive user,
as well as various programs, receiving failure responses
in the average use case.

Therefore we parse more of the request header fields and
specifically read forward through the QNAME field to find
QTYPE and QCLASS.

We check whether the QDCOUNT field is at least 1 before
replying.

We also introduce some basic bounds checks so we don't read
past the end of the received data (even though this would be
safe, given the size of our buffer), and, more importantly,
avoid reading backwards past the start of our buffer, which
was previously possible if one sent a packet of only a few
bytes, e.g., with "echo -n 1 | nc -4u -w0 127.0.0.1 55353".

Note that this is by no means is this a full-fledged or
complete implementation of EDNS, with all the bells and
whistles that would require.  These changes merely attempt
to allow launchdns to respond to queries which contain
data beyond the QNAME withthout always sending NXDOMAIN,
as happens now.

This should permit launchdns to function as an older DNS
responder which simply ignores unknown OPT-type RRs in
queries, so that a basic "dig foo.localhost" succeeds again.