initialize sigaction struct to avoid SA_RESTART
The timeout.bats test would fail intermittently on
Linux, in particular, because the sa sigaction struct
was not initialized (cleared to zeros), so the
sa.sa_flags field might sometimes happen to have
its SA_RESTART flag set (along with other random
flags, of course).

When the SA_RESTART flag was set, and the SIGALRM
signal was received (i.e., when any timeout value
specified on the command line with -t, such as in
the timeout.bats test) while the recvfrom() call
was blocking, the recvfrom() call would be
automatically restarted instead of returning with
errno set to EINTR, thus preventing the main loop
from exiting upon the reciept of the timeout alarm.